+++ localization/graph_localizer/include/graph_localizer/graph_initializer_params.h
@@ -22,6 +22,7 @@
 #include <imu_integration/imu_filter_params.h>
 
 #include <gtsam/geometry/Pose3.h>
+#include <gtsam/base/Vector.h>
 
 #include <string>
 
@@ -29,7 +30,7 @@ namespace graph_localizer {
 // TODO(rsoussan): Clean this up, only use what is needed from imu integration params
 struct GraphInitializerParams : public imu_integration::LatestImuIntegratorParams {
   gtsam::Pose3 global_T_body_start;
-  gtsam::Velocity3 global_V_body_start;
+  gtsam::Vector3 global_V_body_start;
   std::string imu_bias_filename;
   int num_bias_estimation_measurements;
 };
